mocked all API calls for Layout story + lifted mock out of component folder
