+++ fixtures/mostViewed.ts
@@ -278,3 +278,24 @@ export const mockMostShared = {
     pillar: 'opinion',
     ageWarning: '1 year old',
 };
+
+export const responseWithTwoTabs = {
+    tabs: [mockTab1, mockTab2],
+    mostCommented: mockMostCommented,
+    mostShared: mockMostShared,
+};
+
+export const responseWithOneTab = {
+    tabs: [mockTab1],
+    mostCommented: mockMostCommented,
+    mostShared: mockMostShared,
+};
+
+export const responseWithMissingImage = {
+    tabs: [mockTab1],
+    mostCommented: mockMostCommented,
+    mostShared: {
+        ...mockMostShared,
+        image: null,
+    },
+};
