--- src/web/components/MostViewed/MostViewedRight/MostViewedRight.mocks.ts
@@ -1,5 +1,0 @@
-import { mockTab1 } from '@root/fixtures/mostViewed';
-
-export const response = {
-    data: mockTab1,
-};
